add AbstractActionDAO to extend Reply and DeliveryStatus
